+++ test/index.test.ts
@@ -30,9 +30,51 @@ describe("bindl", () => {
     expect(result.exitCode).toBe(0);
 
     expect(shell.test("-f", "./binaries/linux/x64/shellcheck")).toBeTruthy();
+    expect(shell.test("-f", "./binaries/linux/arm/shellcheck")).toBeTruthy();
+    expect(shell.test("-f", "./binaries/linux/arm64/shellcheck")).toBeTruthy();
     expect(shell.test("-f", "./binaries/darwin/x64/shellcheck")).toBeTruthy();
     expect(
       shell.test("-f", "./binaries/win32/x64/shellcheck.exe")
     ).toBeTruthy();
+    expect(
+      shell.test("-f", "./binaries/win32/ia32/shellcheck.exe")
+    ).toBeTruthy();
+    expect(shell.test("-f", "./binaries/win32/ia32/LICENSE.txt")).toBeTruthy();
+    expect(shell.test("-f", "./binaries/win32/ia32/README.txt")).toBeTruthy();
+  });
+
+  it("downloads only linux-arm64 binary when npm_config_arch is set", async () => {
+    const result = await execa.command(`${bin} --config ${configPath}`, {
+      // eslint-disable-next-line camelcase
+      env: { npm_config_arch: "x64" },
+    });
+    expect(result.stdout).toContain(`downloading and extracting`);
+    expect(result.exitCode).toBe(0);
+
+    if (process.platform === "linux") {
+      expect(shell.test("-f", "./binaries/linux/x64/shellcheck")).toBeTruthy();
+    } else if (process.platform === "darwin") {
+      expect(shell.test("-f", "./binaries/darwin/x64/shellcheck")).toBeTruthy();
+    } else if (process.platform === "win32") {
+      expect(
+        shell.test("-f", "./binaries/win32/x64/shellcheck.exe")
+      ).toBeTruthy();
+    }
+    expect(shell.test("-f", "./binaries/linux/arm/shellcheck")).toBeFalsy();
+    expect(shell.test("-f", "./binaries/linux/arm64/shellcheck")).toBeFalsy();
+    expect(shell.test("-f", "./binaries/win32/x64/shellcheck.exe")).toBeFalsy();
+    expect(
+      shell.test("-f", "./binaries/win32/ia32/shellcheck.exe")
+    ).toBeFalsy();
+    expect(shell.test("-f", "./binaries/win32/ia32/LICENSE.txt")).toBeFalsy();
+    expect(shell.test("-f", "./binaries/win32/ia32/README.txt")).toBeFalsy();
+  });
+
+  it("downloads nothing when BINDL_SKIP is set", async () => {
+    const result = await execa.command(`${bin} --config ${configPath}`, {
+      env: { BINDL_SKIP: "true" },
+    });
+    expect(result.stdout).toContain(`Skipping`);
+    expect(result.exitCode).toBe(0);
   });
 });