+++ test/res/bindl.config.js
@@ -1,11 +1,37 @@
-const version = "v0.7.1";
+const version = "v0.8.0";
+const shellcheckReleaseUrl = `https://github.com/koalaman/shellcheck/releases/download/${version}/shellcheck-${version}`;
 
 module.exports = {
+  decompressPlugins: [
+    "@felipecrs/decompress-tarxz"
+  ],
   binaries: [
     {
       platform: "linux",
       arch: "x64",
-      url: `https://github.com/koalaman/shellcheck/releases/download/${version}/shellcheck-${version}.linux.x86_64.tar.xz`,
+      url: `${shellcheckReleaseUrl}.linux.x86_64.tar.xz`,
+      files: [
+        {
+          source: `shellcheck-${version}/shellcheck`,
+          target: "shellcheck",
+        },
+      ],
+    },
+    {
+      platform: "linux",
+      arch: "arm",
+      url: `${shellcheckReleaseUrl}.linux.armv6hf.tar.xz`,
+      files: [
+        {
+          source: `shellcheck-${version}/shellcheck`,
+          target: "shellcheck",
+        },
+      ],
+    },
+    {
+      platform: "linux",
+      arch: "arm64",
+      url: `${shellcheckReleaseUrl}.linux.aarch64.tar.xz`,
       files: [
         {
           source: `shellcheck-${version}/shellcheck`,
@@ -16,7 +42,7 @@ module.exports = {
     {
       platform: "darwin",
       arch: "x64",
-      url: `https://github.com/koalaman/shellcheck/releases/download/${version}/shellcheck-${version}.darwin.x86_64.tar.xz`,
+      url: `${shellcheckReleaseUrl}.darwin.x86_64.tar.xz`,
       files: [
         {
           source: `shellcheck-${version}/shellcheck`,
@@ -26,11 +52,14 @@ module.exports = {
     },
     {
       platform: "win32",
+      arch: "ia32",
+      url: `${shellcheckReleaseUrl}.zip`,
+    },
+    {
+      platform: "win32",
       arch: "x64",
-      url: `https://github.com/koalaman/shellcheck/releases/download/${version}/shellcheck-${version}.zip`,
-      files: [
-        { source: `shellcheck-${version}.exe`, target: "shellcheck.exe" },
-      ],
+      url: `${shellcheckReleaseUrl}.zip`,
+      files: [{ source: `shellcheck.exe`, target: "shellcheck.exe" }],
     },
   ],
 };